+++ editor_button/tmpl/default.php
@@ -24,7 +24,7 @@ $document->addScript( JUri::root() .'/plugins/editors-xtd/pdfviewer/assets/pdfvi
 
 // get all published jdownloads files
 $db = JFactory::getDbo();
-$query = "SELECT id, title FROM #__jdownloads_files WHERE published = 1  ORDER BY publish_up desc";
+$query = "SELECT id, title FROM #__jdownloads_files WHERE published = 1 ORDER BY publish_up desc";
 $db->setQuery($query);
 $fields = $db->loadAssocList();
 
@@ -36,17 +36,82 @@ foreach ($fields as $field) {
 }
 $dropdown .= '</select>';
 
-?>
 
+// Get plugin 'my_plugin' of plugin type 'my_plugin_type'
+$plugin = JPluginHelper::getPlugin('editors-xtd', 'pdfviewer');
+
+
+// Check if plugin is enabled
+if ($plugin)
+{
+    // Get plugin params
+    $pluginParams = new JRegistry($plugin->params);
+
+
+	//select default in viewer dropdown
+	 $paramviewer = $pluginParams->get('viewer');
+	$selectpdfjs = '';
+	$selectpdfimage = '';	
+	switch  ($paramviewer) {
+		case "pdfjs":
+			$selectpdfjs = 'selected';		
+		break;
+	
+		case "pdfimage":
+			$selectpdfimage = 'selected';
+		break;
+		
+		default:
+			$selectpdfjs = 'selected';	
+		break;
+
+	}
+		
+	
+	//select default in style dropdown
+	$paramstyle = $pluginParams->get('style');
+	$selectembed = '';
+	$selectpopup = '';	
+	$selectnew = '';
+
+	$setwidth = '';
+	$setheight = '';
+	
+	switch ($paramstyle) {
+		case "embed":
+			$selectembed = 'selected';
+			$setwidth = 'value="'. $pluginParams->get('embedwidth') .'"';
+			$setheight = 'value="'. $pluginParams->get('embedheight') .'"';
+		break;
+	
+		case "popup":
+			$selectpopup = 'selected';
+			$setwidth = 'value="'. $pluginParams->get('popupwidth') .'"';
+			$setheight = 'value="'. $pluginParams->get('popupheight') .'"';
+		break;
+		
+		case "new":
+			$selectnew = 'selected';
+		break;
+		
+		default:
+			
+			$selectpopup = 'selected';	
+		break;
+	}
+		
+
+}
+?>
 
 
 <div class="container-popup">
 	<form class="form-horizontal">
-	
-			<input type="radio" id="jdownloadsid_radio" name="filetype" value="jdownloadsid" onchange="filesettings()">
-			<label for="jdownloads">Jdownloads</label>
-			<input type="radio" id="file_radio" name="filetype" value="file" onchange="filesettings()">
-			<label for="file">external pdf</label>
+			<div class="form-check form-check-inline">
+			<input type="radio" id="jdownloadsid_radio" name="filetype" value="jdownloadsid" onchange="filesettings()"><label for="jdownloads">Jdownloads</label>
+			</div>
+			<div class="form-check form-check-inline">
+			<input type="radio" id="file_radio" name="filetype" value="file" onchange="filesettings()"><label for="file">external pdf</label>
 						
 		<div id="form_div" style="display: none;">
 		<table style="width:100%" >
@@ -54,22 +119,11 @@ $dropdown .= '</select>';
 			<tr>
 				<td valign=top>
 				
-						
-
-						
-
-				
 		
 						<div id="jdownloadsid_div">
 						<label for="jdownloadsid">jDownloadsid</label>
 						<select id="jdownloadsid" name="jdownloadsid" onchange="filesettings()">
-						<?php
-						echo $dropdown;
-						?>
-						
-					<!--	
-						<label for="jdownloadsid">jDownloadsid</label>	 <input label="jDownloadsid" type="text" id="jdownloadsid" name="jdownloadsid" min="1" style="width:60px" onchange="filesettings()">
-						 -->
+							<?php	echo $dropdown;					?>
 						 </div>
 						
 						<div id="file_div">
@@ -79,9 +133,9 @@ $dropdown .= '</select>';
 						<div id="viewer_div">
 						<label for="viewer">Viewer</label>	
 						<select id="viewer" name="viewer" onchange="viewersettings()">
-							  <option value="default">default</option>
-							  <option value="pdfjs">pdfjs</option>
-							  <option value="pdfimage">pdfimage</option>
+							<!--  <option value="default">default</option> -->
+							  <option value="pdfjs" <?php echo $selectpdfjs ; ?>>pdfjs</option>
+							  <option value="pdfimage" <?php echo $selectpdfimage ; ?>>pdfimage</option>
 							</select>
 						<br><br>
 						</div>
@@ -90,20 +144,20 @@ $dropdown .= '</select>';
 						
 						<label for="style">Style</label>
 						<select id="style" name="style" onchange="stylesettings()"  >
-							  <option value="default">default</option>
-							  <option value="embed">embed</option>
-							  <option value="popup">popup</option>
-							  <option value="new">new</option>
+							 <!-- <option value="default">default</option> -->
+							<option value="embed" <?php echo $selectembed ; ?>>embed</option>
+							<option value="popup" <?php echo $selectpopup ; ?>>popup</option>
+							<option value="new" <?php echo $selectnew ; ?>>new</option>
 							</select>
 						<br><br>
 						
 						<div id="sizesettings_div" >
-						<label for="height">Height</label>	
-							<input label="height" type="text" id="height" name="height" min="0" style="width:40px" >
-						<br>
-						
+													
 						<label for="width">Width</label> 
-							<input label="width" type="text" id="width" name="width" style="width:40px"  >
+							<input label="width" type="text" id="width" name="width" style="width:40px" <?php echo $setwidth; ?>  >
+							<br>
+						<label for="height">Height</label>	
+							<input label="height" type="text" id="height" name="height" min="0" style="width:40px" <?php echo $setheight; ?> >							
 						</div>
 						
 
@@ -115,9 +169,9 @@ $dropdown .= '</select>';
 						</div>
 						<div id="pagenumber_div">
 							<label for="pagenumber">Pagenumber</label>	 <input label="page" type="number" id="page" name="page" min="0" style="width:60px" >
-							<br><br>
+							
 						</div>
-						<br><br>
+						<br>
 						<div id="linktext_div">
 						<label for="Linktext">Linktext</label>	 <input label="linktext" type="text" id="linktext" name="search" >
 						</div>
@@ -140,8 +194,7 @@ $dropdown .= '</select>';
 <script>
 
 	function filesettings() {
-		var file = document.getElementById("jdownloadsid_radio").checked;
-		 
+		var file = document.getElementById("jdownloadsid_radio").checked; 
 
 		if (file ==true ) { //jdownloads file
 			document.getElementById("file_div").style.display = "none";
@@ -149,6 +202,11 @@ $dropdown .= '</select>';
 			document.getElementById("viewer_div").style.display = "block";
 					
 			document.getElementById("form_div").style.display = "block";
+			
+			// set filename as default linktext		
+			var select = document.getElementById('jdownloadsid');
+			var value = select.options[select.selectedIndex].text;			
+			document.getElementById("linktext").value = value;
 							
 
 		} else { // external file
@@ -157,6 +215,8 @@ $dropdown .= '</select>';
 			document.getElementById("viewer_div").style.display = "none";
 			
 			document.getElementById("form_div").style.display = "block";
+			
+			document.getElementById("linktext").value = "";
 
 		}
 	
@@ -184,7 +244,10 @@ $dropdown .= '</select>';
 
 	function stylesettings() {
 		var style = document.getElementById("style").value;
-		 
+		
+		
+		document.getElementById("width").value = ""; 
+		document.getElementById("height").value = ""; 
 		 
 		if (style == 'embed' || style == 'popup') {
 			document.getElementById("sizesettings_div").style.display = "block";